instructor bio done

+++ Util/deleteFile.js
@@ -1,19 +1,30 @@
 const fs = require('fs');
 
 const deleteSingleFile = (filePath) => {
-    fs.unlink(filePath, (err) => {
-        if (err) {
-            throw (err);
+    if (filePath) {
+        // console.log(fs.existsSync(filePath));
+        if (fs.existsSync(filePath)) {
+            fs.unlink(filePath, (err) => {
+                if (err) {
+                    throw (err);
+                }
+            })
         }
-    })
+    }
 }
 
 const deleteMultiFile = (filePath) => {
-    filePath.map(path => fs.unlink(path, (err) => {
-        if (err) {
-            throw (err);
+    filePath.map(path => {
+        if (path) {
+            if (fs.existsSync(path)) {
+                fs.unlink(path, (err) => {
+                    if (err) {
+                        throw (err);
+                    }
+                });
+            }
         }
-    }))
+    })
 }
 
 module.exports = {
